+++ src/crafting.cpp
@@ -792,7 +792,7 @@ recipe* game::select_crafting_recipe()
                     for (unsigned j = 0; j < current[line]->tools[i].size(); j++)
                     {
                         itype_id type = current[line]->tools[i][j].type;
-                        int charges = current[line]->tools[i][j].count;
+                        long charges = current[line]->tools[i][j].count;
                         nc_color toolcol = has_one ? c_dkgray : c_red;
 
                         if (current[line]->tools[i][j].available == 0)
@@ -1595,7 +1595,7 @@ void game::consume_tools(player *p, std::vector<component> tools, bool force_ava
   }
   itype_id type = tools[i].type;
   if (tools[i].count > 0) {
-   int count = tools[i].count;
+   long count = tools[i].count;
    if (p->has_charges(type, count))
     player_has.push_back(tools[i]);
    if (map_inv.has_charges(type, count))
