+++ src/vehicle.h
@@ -505,7 +505,7 @@ public:
     void fire_turret (int p, bool burst = true);
 
     // internal procedure of turret firing
-    bool fire_turret_internal (int p, it_gun &gun, it_ammo &ammo, int charges,
+    bool fire_turret_internal (int p, it_gun &gun, it_ammo &ammo, long charges,
                                const std::string &firing_sound = "");
 
     // opens/closes doors or multipart doors
